ajustes

+++ app/src/main/java/br/com/freelas/app/mobile/raspe/mania/raspemania/model/entidade/Produto.java
@@ -7,11 +7,11 @@ import br.com.freelas.app.mobile.raspe.mania.raspemania.model.BaseModel;
 public class Produto extends BaseModel {
 
     public String nome;
-    public long valor;
+    public float valor;
 
     public Produto(){}
 
-    public Produto(String nome, long valor){
+    public Produto(String nome, float valor){
         this.nome = nome;
         this.valor = valor;
     }
